requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pip freeze > requirements.txt